[BEAM-7144] Fix for rescaling problem on Flink >= 1.6
After a decent amount of debugging of the checkpoint/restore/rescale code, I
managed to find the culprit. The issue here is the lazy state creation that Beam
uses by default. State is created in Flink's state backend upon first "binding"
of the state. Binding is only performed when processElement is called. Now, it
can happen that an operator does not receive elements before checkpointed. In
this case the state won't be initialized.

Depending on the maxParallelism, this can lead to the state being partitioned
such that an operator attempts to restore from a keygroup that did not have the
state in its meta data. However, the actual data ends up containing the state
which looks like a bug in the Flink key group partitioner.

The workaround is to ensure that all user state is registered before
checkpointing is first attempted. This resolves all issues with rescaling Beam
jobs.